Fix rejection of arrays with eight or nine elements.
In general, LLVM's parsing of decimal integers to `APInt`s forms an
`APInt` that is 4n bits wide, where n is the length of the integer, and
the `isNegative` check only checks the high bit.

In this case, we form an `APInt` that is four bits wide, with the high bit
set, which we reject because we think it's "negative". These two array
lengths are the only ones where this happens -- if the decimal integer
value is two characters long, we form an `APInt` that is eight bits wide
but holds a value < 100, so the high bit is never set, and the same
applies for longer integers too.

An `IntegerLiteral` is never negative, so we don't need the `isNegative`
check, and in fact it only detects the n=8 and n=9 cases.