Clean up interface of node block stack after recent changes.
Peek() no longer returns a useful value. Remove, and rename PeekForAdd()
to PeekOrAdd() since it's no longer used to get a block ID to add
elements into. PushForAdd() was unused and is removed. Add
NodeBlockStack::AddNode to combine the operations of adding a node and
inserting it into a block.