Place the computation of the destination of an initialization into the return slot.
Fix a bug where we would perform the computation of the return location
after we have already used it in some cases, leading to assertion
failures during lowering. Instead, accumulate a sequence of instructions
to compute the return location in a temporary block, and overwrite the
return slot with those instructions when we perform initialization.

StubReference is replaced by a more general SpliceBlock node, that takes
a code block and a result value, executes the instructions in the block,
and produces the result. This is used in the uncommon case where more
than one instruction is required to compute the return slot, which can
happen if we need to first emit a temporary and then index into it, or
if we need to perform multiple levels of indexing before we reach an
entity to initialize.