Rename //toolchain/sem_ir:file_test -> :yaml_test.
While this test is testing //toolchain/sem_ir:file, the interesting
thing about it is that it's testing the YAML output from File printing.
Calling this test `:file_test` is also deeply confusing because that's
also the name we give to all of our tests that are testing the files in
`testdata/`, which is not what this test does.